requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cssselect==1.1.0
distlib==0.3.0
distro==1.4.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jieba3k==0.35.1
joblib==0.17.0
lockfile==0.12.2
lxml==4.6.1
msgpack==0.6.2
newspaper3k==0.2.8
nltk==3.5
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
PyYAML==5.3.1
regex==2020.11.13
requests==2.22.0
requests-file==1.5.1
retrying==1.3.3
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==3.0.2
tqdm==4.52.0
urllib3==1.25.8
webencodings==0.5.1
wget==3.2
